+++ crates/release_plz_core/src/lib.rs
@@ -7,11 +7,13 @@ mod release_pr;
 mod tmp_repo;
 mod update;
 mod version;
+mod github_client;
 
 pub use changelog::*;
 pub use download::read_package;
 pub use next_ver::*;
 pub use release_pr::*;
 pub use update::*;
+pub use github_client::GitHub;
 
 pub const CARGO_TOML: &str = "Cargo.toml";
