+++ crates/release_plz_core/src/release_pr.rs
@@ -5,43 +5,18 @@ use chrono::SecondsFormat;
 use git_cmd::Repo;
 
 use anyhow::{anyhow, Context};
-use octocrab::OctocrabBuilder;
-use secrecy::{ExposeSecret, SecretString};
-use tracing::{info, instrument, Span};
-use url::Url;
+use tracing::instrument;
 
-use crate::{copy_to_temp_dir, update, UpdateRequest, UpdateResult, CARGO_TOML};
+use crate::{
+    copy_to_temp_dir,
+    github_client::{GitHub, GitHubClient, Pr},
+    update, UpdateRequest, UpdateResult, CARGO_TOML,
+};
 
 #[derive(Debug)]
 pub struct ReleasePrRequest {
     pub github: GitHub,
     pub update_request: UpdateRequest,
-}
-
-#[derive(Debug)]
-pub struct GitHub {
-    pub owner: String,
-    pub repo: String,
-    pub token: SecretString,
-    base_url: Option<Url>,
-}
-
-impl GitHub {
-    pub fn new(owner: String, repo: String, token: SecretString) -> Self {
-        Self {
-            owner,
-            repo,
-            token,
-            base_url: None,
-        }
-    }
-
-    pub fn with_base_url(self, base_url: Url) -> Self {
-        Self {
-            base_url: Some(base_url),
-            ..self
-        }
-    }
 }
 
 /// Open a pull request with the next packages versions of a local rust project
@@ -66,24 +41,20 @@ pub async fn release_pr(input: &ReleasePrRequest) -> anyhow::Result<()> {
         .set_local_manifest(local_manifest)
         .context("can't find temporary project")?;
     let (packages_to_update, _repository) = update(&new_update_request)?;
+    let gh_client = GitHubClient::new(&input.github)?;
+    gh_client.close_other_prs()?;
     if !packages_to_update.is_empty() {
         let repo = Repo::new(new_manifest_dir)?;
-        let pr = Pr::new(&packages_to_update);
+        let pr = Pr::from(packages_to_update.as_ref());
         create_release_branch(&repo, &pr.branch)?;
-        open_pr(&pr, &input.github).await?;
+        gh_client.open_pr(&pr).await?;
     }
 
     Ok(())
 }
 
-#[derive(Debug)]
-struct Pr {
-    branch: String,
-    title: String,
-}
-
-impl Pr {
-    fn new(packages_to_update: &[(Package, UpdateResult)]) -> Self {
+impl From<&[(Package, UpdateResult)]> for Pr {
+    fn from(packages_to_update: &[(Package, UpdateResult)]) -> Self {
         Self {
             branch: release_branch(),
             title: pr_title(packages_to_update),
@@ -107,51 +78,6 @@ fn pr_title(packages_to_update: &[(Package, UpdateResult)]) -> String {
     } else {
         "chore: release".to_string()
     }
-}
-
-#[instrument(
-    fields(
-        default_branch = tracing::field::Empty,
-    )
-)]
-async fn open_pr(pr: &Pr, github: &GitHub) -> anyhow::Result<()> {
-    let mut octocrab_builder =
-        OctocrabBuilder::new().personal_token(github.token.expose_secret().clone());
-
-    if let Some(base_url) = &github.base_url {
-        octocrab_builder = octocrab_builder
-            .base_url(base_url.clone())
-            .context("Invalid GitHub base url")?;
-    }
-
-    let client = octocrab_builder
-        .build()
-        .context("Failed to build GitHub client")?;
-
-    let default_branch = client
-        .repos(&github.owner, &github.repo)
-        .get()
-        .await
-        .context(format!(
-            "failed to retrieve GitHub repository {}/{}",
-            github.owner, github.repo
-        ))?
-        .default_branch
-        .context("failed to retrieve default branch")?;
-    Span::current().record("default_branch", &default_branch.as_str());
-
-    let pr = client
-        .pulls(&github.owner, &github.repo)
-        .create(&pr.title, &pr.branch, default_branch)
-        .body("release-plz automatic bot")
-        .send()
-        .await?;
-
-    if let Some(url) = pr.html_url {
-        info!("opened pr: {}", url);
-    }
-
-    Ok(())
 }
 
 fn create_release_branch(repository: &Repo, release_branch: &str) -> anyhow::Result<()> {
